requirements.txt
tensorflow>=2.15.0,<3.0.0
gradio>=6.7.0,<7.0.0
numpy>=1.26.0,<2.0.0
Pillow>=10.0.0,<11.0.0
Pillow>=10.2.0
